DZIAŁA HURRAA

+++ app/src/main/java/MainActivity/Medication.kt
@@ -2,7 +2,7 @@ package MainActivity
 
 data class Medication(
     val drugName: String,
-    val dosage: Int,
+    val dosage: Double,
     val unit: String,
     val time: String
 )
